requirements.txt
Django == 3.2
psycopg2 >= 2.8
djangorestframework == 3.13.1
factory-boy == 3.2.1
django-extensions == 3.2.1
pydantic == 1.10.2
djantic == 0.7.0
django-ninja == 0.19.1